+++ packages/core-modules/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @eclesia/core-modules-pg
 
+## 2.9.3
+
+### Patch Changes
+
+- Fix events for comet38
+
 ## 2.9.2
 
 ### Patch Changes